Move around benches files
