+++ Makefile
@@ -6,7 +6,7 @@ test:
 	cargo test --release --features=lp
 	# don't run examples in proof-production mode
 	cargo test --release --features "test-explanations"
-	
+
 
 .PHONY: nits
 nits:
@@ -24,3 +24,28 @@ nits:
 .PHONY: docs
 docs:
 	RUSTDOCFLAGS="--cfg docsrs" cargo +nightly doc --all-features --open
+
+
+
+math.csv:
+	EGG_BENCH_CSV=math.csv cargo test --test math --release -- --nocapture --test --test-threads=1
+
+lambda.csv:
+	EGG_BENCH_CSV=lambda.csv cargo test --test lambda --release -- --nocapture --test --test-threads=1
+
+.PHONY: existing-bench
+existing-bench: math.csv lambda.csv
+
+.PHONY: clean-bench
+clean-bench:
+	rm math.csv lambda.csv profile.json
+
+.PHONY: bench
+bench:
+	cargo build --profile test && cargo bench
+
+profile.json:
+	cargo build --profile test && samply record cargo bench
+
+.PHONY: profile
+profile: profile.json
