+++ benches/simple.rs
@@ -4,99 +4,10 @@ use rayon::prelude::*;
 mod definitions;
 use definitions::simple;
 
+mod schedulers;
+use schedulers::schedulers::*;
+
 use criterion::{criterion_group, criterion_main, Criterion, BenchmarkId};
-
-pub struct SerialRewriteScheduler;
-impl<L: Language, N: Analysis<L>> RewriteScheduler<L, N> for SerialRewriteScheduler {
-    fn search_rewrites<'a>(
-        &mut self,
-        iteration: usize,
-        egraph: &EGraph<L, N>,
-        rewrites: &[&'a Rewrite<L, N>],
-        limits: &RunnerLimits,
-    ) -> RunnerResult<Vec<Vec<SearchMatches<'a, L>>>> {
-        rewrites
-            .iter()
-            .map(|rw| {
-                let ms = rw.search(egraph);
-                limits.check_limits(iteration, egraph)?;
-                Ok(ms)
-            })
-            .collect()
-    }
-}
-
-pub struct ParallelRewriteScheduler;
-impl<L, N> RewriteScheduler<L, N> for ParallelRewriteScheduler
-    where
-    L: Language + Sync + Send,
-    L::Discriminant: Sync + Send,
-    N: Analysis<L> + Sync + Send,
-    N::Data: Sync + Send
-{
-// impl<L: Language + Send + Sync> RewriteScheduler<L, ()> for ParallelRewriteScheduler {
-    fn search_rewrites<'a>(
-        &mut self,
-        iteration: usize,
-        egraph: &EGraph<L, N>,
-        rewrites: &[&'a Rewrite<L, N>],
-        limits: &RunnerLimits,
-    ) -> RunnerResult<Vec<Vec<SearchMatches<'a, L>>>> {
-        // This implementation just ignores the limits
-        // fake `par_map` to enforce Send + Sync, in real life use rayon
-        // fn par_map<T, F, T2>(slice: &[T], f: F) -> Vec<T2>
-        // where
-        //     T: Send + Sync,
-        //     F: Fn(&T) -> T2 + Send + Sync,
-        //     T2: Send + Sync,
-        // {
-        //     slice.iter().map(f).collect()
-        // }
-        // Ok(par_map(rewrites, |rw| rw.search(egraph)))
-
-        rewrites
-            .par_iter()
-            .map(|rw| {
-                let ms = rw.search(egraph);
-                limits.check_limits(iteration, egraph)?;
-                Ok(ms)
-            })
-            .collect() // ::<RunnerResult<Vec<Vec<SearchMatches<'a, L>>>>>()
-
-        // TODO: Note that `Sync + Send` traits were added to both language and
-        //       discriminant. Could this impact correctness?
-    }
-}
-
-
-pub struct RestrictedParallelRewriteScheduler;
-impl<L> RewriteScheduler<L, ()> for RestrictedParallelRewriteScheduler
-    where
-    L: Language + Sync + Send,
-    L::Discriminant: Sync + Send,
-{
-// impl<L: Language + Send + Sync> RewriteScheduler<L, ()> for ParallelRewriteScheduler {
-    fn search_rewrites<'a>(
-        &mut self,
-        iteration: usize,
-        egraph: &EGraph<L, ()>,
-        rewrites: &[&'a Rewrite<L, ()>],
-        limits: &RunnerLimits,
-    ) -> RunnerResult<Vec<Vec<SearchMatches<'a, L>>>> {
-        rewrites
-            .par_iter()
-            .map(|rw| {
-                let ms = rw.search(egraph);
-                limits.check_limits(iteration, egraph)?;
-                Ok(ms)
-            })
-            .collect()
-    }
-}
-
-
-
-
 
 
 fn serial_simplify(s: &str) -> String {
@@ -177,8 +88,6 @@ pub fn comparison_simple_bench(c: &mut Criterion) {
 }
 
 
-
-
 // fn math_serial_simplify_root() {
 //     egg::test::test_runner(
 //         "math_simplify_root",
@@ -206,8 +115,6 @@ pub fn comparison_simple_bench(c: &mut Criterion) {
 //     //    |b| b.iter(math_simplify_factor)
 //     //);
 // }
-
-
 
 
 criterion_group!(benches, comparison_simple_bench);
